Added filter fixtures.
